be/requirements.txt
# requirements.txt
Flask==2.3.2
flask-sqlalchemy==3.0.3
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.2.2
pandas==2.0.1
python-dotenv==1.0.0
pytest==7.3.1